[IMPROVEMENT] Adds log directives for link (project) specific logging in NginX.
Previously NginX would log everything to a single set of access and
error log, which would make is hard to evaluate the logs when looking
for errors.
The purpose of this change is to allow one set of log files for every
link that is created, which should make it a lot easier to evaluate the
logs.