[FIX] Fixes that multidomain support could lead to strange behaviour
When a multidomain value was specified, that consists of the starting
letters of the link name (e.g. value 'tr' for link name 'transfigure'),
this could result in the wrong store view being shown when trying to
access the default store view.
The reason was that the specified value is supposed to be a subdomain
followed by a dot. In the regular expression though, the dot was not
escaped and thus not literal but rather a wildcard.